Module My_printtyped

These extra functions are an extension of Printtyped module from compiler libs. Maybe in new versions of compiler they will be available out of box.

include module type of struct include Printtyped end

val interface : Stdlib.Format.formatter -> Typedtree.signature -> unit

val implementation : Stdlib.Format.formatter -> Typedtree.structure -> unit

val implementation_with_coercion : Stdlib.Format.formatter -> Typedtree.implementation -> unit

New functions

val expr : Stdlib.Format.formatter -> Typedtree.expression -> unit

val me : Stdlib.Format.formatter -> Typedtree.module_expr -> unit

val attrs : Stdlib.Format.formatter -> Ast_helper.attrs -> unit