Module My_untype

These extra functions are an extension of Untypeast module from compiler libs. Maybe in new versions of compiler they will be available out of box.

include module type of struct include Untypeast end

val lident_of_path : Path.t -> Longident.t

type mapper = Untypeast.mapper = {

attribute : mapper -> Typedtree.attribute -> Parsetree.attribute;

attributes : mapper -> Typedtree.attribute list -> Parsetree.attribute list;

binding_op : mapper -> Typedtree.binding_op -> Typedtree.pattern -> Parsetree.binding_op;

case : 'k. mapper -> 'k Typedtree.case -> Parsetree.case;

class_declaration : mapper -> Typedtree.class_declaration -> Parsetree.class_declaration;

class_description : mapper -> Typedtree.class_description -> Parsetree.class_description;

class_expr : mapper -> Typedtree.class_expr -> Parsetree.class_expr;

class_field : mapper -> Typedtree.class_field -> Parsetree.class_field;

class_signature : mapper -> Typedtree.class_signature -> Parsetree.class_signature;

class_structure : mapper -> Typedtree.class_structure -> Parsetree.class_structure;

class_type : mapper -> Typedtree.class_type -> Parsetree.class_type;

class_type_declaration : mapper -> Typedtree.class_type_declaration -> Parsetree.class_type_declaration;

class_type_field : mapper -> Typedtree.class_type_field -> Parsetree.class_type_field;

constructor_declaration : mapper -> Typedtree.constructor_declaration -> Parsetree.constructor_declaration;

expr : mapper -> Typedtree.expression -> Parsetree.expression;

extension_constructor : mapper -> Typedtree.extension_constructor -> Parsetree.extension_constructor;

include_declaration : mapper -> Typedtree.include_declaration -> Parsetree.include_declaration;

include_description : mapper -> Typedtree.include_description -> Parsetree.include_description;

label_declaration : mapper -> Typedtree.label_declaration -> Parsetree.label_declaration;

location : mapper -> Location.t -> Location.t;

module_binding : mapper -> Typedtree.module_binding -> Parsetree.module_binding;

module_declaration : mapper -> Typedtree.module_declaration -> Parsetree.module_declaration;

module_substitution : mapper -> Typedtree.module_substitution -> Parsetree.module_substitution;

module_expr : mapper -> Typedtree.module_expr -> Parsetree.module_expr;

module_type : mapper -> Typedtree.module_type -> Parsetree.module_type;

module_type_declaration : mapper -> Typedtree.module_type_declaration -> Parsetree.module_type_declaration;

package_type : mapper -> Typedtree.package_type -> Parsetree.package_type;

open_declaration : mapper -> Typedtree.open_declaration -> Parsetree.open_declaration;

open_description : mapper -> Typedtree.open_description -> Parsetree.open_description;

pat : 'k. mapper -> 'k Typedtree.general_pattern -> Parsetree.pattern;

row_field : mapper -> Typedtree.row_field -> Parsetree.row_field;

object_field : mapper -> Typedtree.object_field -> Parsetree.object_field;

signature : mapper -> Typedtree.signature -> Parsetree.signature;

signature_item : mapper -> Typedtree.signature_item -> Parsetree.signature_item;

structure : mapper -> Typedtree.structure -> Parsetree.structure;

structure_item : mapper -> Typedtree.structure_item -> Parsetree.structure_item;

typ : mapper -> Typedtree.core_type -> Parsetree.core_type;

type_declaration : mapper -> Typedtree.type_declaration -> Parsetree.type_declaration;

type_extension : mapper -> Typedtree.type_extension -> Parsetree.type_extension;

type_exception : mapper -> Typedtree.type_exception -> Parsetree.type_exception;

type_kind : mapper -> Typedtree.type_kind -> Parsetree.type_kind;

value_binding : mapper -> Typedtree.value_binding -> Parsetree.value_binding;

value_description : mapper -> Typedtree.value_description -> Parsetree.value_description;

with_constraint : mapper -> (Path.t * Longident.t Location.loc * Typedtree.with_constraint) -> Parsetree.with_constraint;

}

val default_mapper : mapper

val untype_structure : ?mapper:mapper -> Typedtree.structure -> Parsetree.structure

val untype_signature : ?mapper:mapper -> Typedtree.signature -> Parsetree.signature

val untype_expression : ?mapper:mapper -> Typedtree.expression -> Parsetree.expression

val untype_pattern : ?mapper:mapper -> 'a Typedtree.general_pattern -> Parsetree.pattern

val constant : Asttypes.constant -> Parsetree.constant

New functions

val expr : ?mapper:mapper -> Typedtree.expression -> Parsetree.expression

val untype_stru_item : Typedtree.structure_item -> Parsetree.structure_item