Module Zanuda_core

module Collected_decls : sig ... end

Global storage for defined names. Is used to detect public but not used library fields.

module Collected_lints : sig ... end

Global storage for found defects.

module Config : sig ... end

module Dune_project : sig ... end

Specialized data type to deserialize the output of 'dune describe'

module LINT : sig ... end

Module types for program tree analyzers

module Lint_filesystem : sig ... end

module Load_dune : sig ... end

module Unused_ML_logger : sig ... end

Collect used values from structure files.

module Utils : sig ... end

Various helper functions.