Module Zanuda_core.Collected_lints

Global storage for found defects.

Mutable. Not thread safe.

Collecting found lints

val clear : unit -> unit

val add : loc:Warnings.loc -> (module LINT.REPORTER) -> unit

val report : unit -> unit

Report found lints

In RdJSONl format. Change Config.t.out_rdjsonl to modify output file name

As plain text to stdout

Collecting type declarations

We use information about type declarations to skip reporting lints in the code generated from a type declaration via `deriving`.

val add_tdecl : Warnings.loc -> unit

val has_tdecl_at : Warnings.loc -> bool

val clear_tdecls : unit -> unit

val process_switches : string -> unit