Module Zanuda_core.LINT

Module types for program tree analyzers

type impl =

| Untyped

| Typed

A type of lint's implementation. Typed and untyped lints inspect OCaml's Parsetree/Typedtree respectively. In Clippy it corresponds to early and late lint passes.

type group =

| Style
(*
Code that should be written in a more idiomatic way
*)

| Correctness
(*
Code that is outright wrong or useless
*)

| Perf
(*
Code that can be written to run faster
*)

| Restriction

| Deprecated

| Pedantic
(*
Lints which are rather strict or might have false positives
*)

| Complexity

| Suspicious

| Nursery
(*
New lints that are still under development
*)

Group of lints. The same as Rust's Clippy

type level =

| Allow

| Warn

| Deny

| Deprecated

Level of lints. The same as Rust's Clippy

type lint_source =

| Camelot
(*
Adopted from Camelot linter
*)

| Clippy
(*
Adopted from Rust's Clippy
*)

| FPCourse
(*
Invented after reviewing Kakadu's student's OCaml homeworks
*)

| Other
(*
The source is not specified
*)

How various lints were invented

module type GENERAL = sig ... end

module type UNTYPED = sig ... end

module type TYPED = sig ... end

module type UNUSED_DECLS = sig ... end

module type REPORTER = sig ... end