Package zanuda 1.0.0

My_printtyped These extra functions are an extension of Printtyped module from compiler libs. Maybe in new versions of compiler they will be available out of box.

My_untype These extra functions are an extension of Untypeast module from compiler libs. Maybe in new versions of compiler they will be available out of box.

Tast_pattern Matching combinators for Typedtree.

Zanuda_core

Package info

authors

Kakadu

changes-files

CHANGES.md

depends

angstrom

base

bisect_ppx

dune

dune-build-info

menhir

ocaml

ocamlformat

odig

odoc

ppx_assert

ppx_blob

ppx_deriving

ppx_expect

ppx_fields_conv

ppx_inline_test

sexplib

stdune

yojson

homepage

https://github.com/Kakadu/zanuda

issues

https://github.com/Kakadu/zanuda/issues

license

LGPL-3.0-only

license-files

LICENSE

maintainers

kakadu@pm.me

readme-files

README.md

repo

git+https://github.com/Kakadu/zanuda.git

tags

lint

test

version

1.0.0